[FEAT/#94] 예외 처리 코드 추가 - deleteReviewLike, getReviewDetail 서비스 예외 처리 추가 - 사용하지 않는 에러 코드 삭제 - 사용하지 않는 파라미터 삭제

+++ src/main/java/com/example/picknwhip_be/domain/review/service/command/ReviewCommandServiceImpl.java
@@ -156,6 +156,13 @@ public class ReviewCommandServiceImpl implements ReviewCommandService {
   @Override
   @Transactional
   public ReviewResDTO.ReviewLikeDTO deleteReviewLike(Long reviewId, Long userId) {
+    if (!reviewRepository.existsById(reviewId)) {
+      throw new ReviewException(ReviewErrorCode.REVIEW_NOT_FOUND);
+    }
+    if (!userRepository.existsById(userId)) {
+      throw new UserException(UserErrorCode.USER_NOT_FOUND);
+    }
+
     reviewLikeRepository.deleteByReviewIdAndUserUserId(reviewId, userId);
     long likeCount = reviewLikeRepository.countByReviewId(reviewId);
     return ReviewConverter.toReviewLikeDTO(reviewId, false, likeCount);
